fix: use a byte instead of a short to write MCEmoji

+++ src/main/java/team/unnamed/emojis/io/writer/MCEmojiWriter.java
@@ -29,7 +29,7 @@ public class MCEmojiWriter implements EmojiWriter {
         dataOutput.write(MCEmojiFormat.VERSION);
 
         // write emoji length
-        dataOutput.writeShort(emojis.size());
+        dataOutput.writeByte(emojis.size());
 
         // write all emojis
         for (Emoji emoji : emojis) {
